Organize preview data and make it filterable per email.
