+++ src/Tickets/Emails/Admin/Preview_Data.php
@@ -17,6 +17,54 @@ namespace TEC\Tickets\Emails\Admin;
  * @package TEC\Tickets\Emails\Admin
  */
 class Preview_Data {
+	/**
+	 * Get default preview data.
+	 *
+	 * @since TBD
+	 *
+	 * @return array The default preview data.
+	 */
+	public static function get_default_preview_data(): array {
+		$current_user = wp_get_current_user();
+		$title        = empty( $current_user->first_name ) ?
+		__( 'Here\'s your ticket!', 'event-tickets' ) :
+		sprintf(
+			// Translators: %s - First name of email recipient.
+			__( 'Here\'s your ticket, %s!', 'event-tickets' ),
+			$current_user->first_name
+		);
+
+		return [
+			'title'      => $title,
+			'heading'    => $title,
+			'is_preview' => true,
+			'tickets'    => [
+				[
+					'ticket_id'         => '1234',
+					'ticket_name'       => esc_html__( 'General Admission', 'event-tickets' ),
+					'holder_name'       => $current_user->first_name . ' ' . $current_user->last_name,
+					'holder_first_name' => $current_user->first_name,
+					'holder_last_name'  => $current_user->last_name,
+					'security_code'     => '17e4a14cec',
+					// @todo @juanfra @codingmusician @rafsuntaskin: These should come from TEC.
+					'event' => [
+						'title'          => esc_html__( 'Rebirth Brass Band', 'event-tickets' ),
+						'description'    => '<h4>Additional Information</h4><p>Age Restriction: 18+<br>Door Time: 8:00PM<br>Event Time: 9:00PM</p>',
+						'date'           => esc_html__( 'September 22 @ 7:00 pm - 11:00 pm', 'event-tickets' ),
+						'image_url'      => esc_url( plugins_url( '/event-tickets/src/resources/images/example-event-image.png' ) ),
+						'venue'          => [
+							'name'       => esc_html__( 'Saturn', 'event-tickets' ),
+							'address1'   => esc_html__( '200 41st Street South', 'event-tickets' ),
+							'address2'   => esc_html__( 'Birmingham, AL, 35222', 'event-tickets' ),
+							'phone'      => esc_html__( '(987) 654-3210', 'event-tickets' ),
+							'website'    => esc_url( get_site_url() ),
+						]
+					],
+
+				],
+			],
+		];
+	}
 
 	/**
 	 * Get Order preview data.
@@ -34,6 +82,7 @@ class Preview_Data {
 			'provider'   => 'Stripe',
 			'status'     => 'success',
 			'post_title' => __( 'Black Midi with Special Guests Chat Pile and Apprehend', 'event-tickets' ),
+			// @todo @codingmusician: We will need to make this work with the currency settings selected for Tickets Commerce.
 			'total'      => '$100.00',
 			'purchaser'  => [
 				'name'  => __( 'John Doe', 'event-tickets' ),
@@ -104,6 +153,7 @@ class Preview_Data {
 			[
 				'title'    => __( 'General Admission', 'event-tickets' ),
 				'quantity' => 2,
+				// @todo @codingmusician: We will need to make this work with the currency settings selected for Tickets Commerce.
 				'price'    => '$50.00',
 			],
 		];
