+++ src/Tickets/Emails/Email/Failed_Order.php
@@ -176,24 +176,26 @@ class Failed_Order extends \TEC\Tickets\Emails\Email_Abstract {
 	}
 
 	/**
-	 * Get preview context for email.
+	 * Get default preview context for email.
 	 *
-	 * @since 5.5.10
+	 * @since TBD
 	 *
 	 * @param array $args The arguments.
 	 * @return array $args The modified arguments
 	 */
-	public function get_preview_context( $args = [] ): array {
+	public function get_default_preview_context( $args = [] ): array {
 		$defaults = [
 			'email'              => $this,
 			'is_preview'         => true,
 			'title'              => $this->get_heading(),
 			'heading'            => $this->get_heading(),
 			'additional_content' => $this->get_additional_content(),
-			'order'              => Preview_Data::get_order( [ 
-				'status'        => 'failed',
-				'error_message' => __( 'Stripe payment processing was unsuccessful.', 'event-tickets' ),
-			] ),
+			'order'              => Preview_Data::get_order(
+				[
+					'status'        => 'failed',
+					'error_message' => __( 'Stripe payment processing was unsuccessful.', 'event-tickets' ),
+				]
+			),
 			'tickets'            => Preview_Data::get_tickets(),
 		];
